api/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
mysql-connector-python==8.2.0
dashscope>=1.23.3
requests>=2.31.0
Werkzeug==2.3.7
python-dotenv>=1.0.0 